refactor(client): :zap: moved non-route folders out of `/app`
