+++ packages/client/assets/avatar.png
