+++ packages/client/components/tabs/diagnostics.tsx
