+++ packages/client/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@client/app/lib/utils"
+import { cn } from "@client/lib/utils";
 
 const Card = React.forwardRef<
   HTMLDivElement,
