+++ packages/client/components/ui/label.tsx
@@ -3,8 +3,8 @@
 import * as React from "react"
 import * as LabelPrimitive from "@radix-ui/react-label"
 import { cva, type VariantProps } from "class-variance-authority"
-
-import { cn } from "@client/app/lib/utils"
+//
+import { cn } from "@client/lib/utils"
 
 const labelVariants = cva(
   "text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
